Fixes back and forth migrations

+++ harvester/metadata/migrations/0008_new_fields.py
@@ -8,54 +8,57 @@ from django.db import migrations, models
 from metadata.models import MetadataField, MetadataTranslation, MetadataValue
 
 
+LEGACY_FIELDS = [
+    "publisher_date", "study_vocabulary", "learning_material_disciplines_normalized", "language.keyword",
+    "copyright.keyword", "technical_type",
+]
+
+NEW_FIELDS = {
+    "study_vocabulary.keyword": {
+        "is_manual": False,
+        "is_hidden": False,
+        "entity": "products:default",
+    },
+    "disciplines_normalized.keyword": {
+        "is_manual": False,
+        "is_hidden": False,
+        "entity": "products:default",
+    },
+    "published_at": {
+        "is_manual": True,
+        "is_hidden": True,
+        "entity": "products:default",
+    },
+    "modified_at": {
+        "is_manual": True,
+        "is_hidden": True,
+        "entity": "products:default",
+    },
+    "language": {
+        "is_manual": False,
+        "is_hidden": False,
+        "entity": "products:default",
+    },
+    "licenses": {
+        "is_manual": False,
+        "is_hidden": False,
+        "entity": "products:default",
+    },
+    "technical_types": {
+        "is_manual": False,
+        "is_hidden": False,
+        "entity": "products:default",
+    }
+}
+
+
 def migrate_metadata_field_to_entities(apps, schema_editor):
     # Skipping this migration for unit test environments
     if not MetadataField.objects.all().exists():
         return
 
-    legacy_fields = [
-        "publisher_date", "study_vocabulary", "learning_material_disciplines_normalized", "language.keyword",
-        "copyright.keyword", "technical_type",
-    ]
-    MetadataField.objects.filter(name__in=legacy_fields).update(entity="products:multilingual-indices")
+    MetadataField.objects.filter(name__in=LEGACY_FIELDS).update(entity="products:multilingual-indices")
 
-    new_fields = {
-        "study_vocabulary.keyword": {
-            "is_manual": False,
-            "is_hidden": False,
-            "entity": "products:default",
-        },
-        "disciplines_normalized.keyword": {
-            "is_manual": False,
-            "is_hidden": False,
-            "entity": "products:default",
-        },
-        "published_at": {
-            "is_manual": True,
-            "is_hidden": True,
-            "entity": "products:default",
-        },
-        "modified_at": {
-            "is_manual": True,
-            "is_hidden": True,
-            "entity": "products:default",
-        },
-        "language": {
-            "is_manual": False,
-            "is_hidden": False,
-            "entity": "products:default",
-        },
-        "licenses": {
-            "is_manual": False,
-            "is_hidden": False,
-            "entity": "products:default",
-        },
-        "technical_types": {
-            "is_manual": False,
-            "is_hidden": False,
-            "entity": "products:default",
-        }
-    }
     edusources_only_fields = ["study_vocabulary.keyword", "disciplines_normalized.keyword"]
     existing_value_translation_fields = {
         "study_vocabulary.keyword": "study_vocabulary",
@@ -64,7 +67,7 @@ def migrate_metadata_field_to_entities(apps, schema_editor):
         "licenses": "copyright.keyword",
         "technical_types": "technical_type",
     }
-    for field_name, defaults in new_fields.items():
+    for field_name, defaults in NEW_FIELDS.items():
         if settings.PLATFORM.value == "publinova" and field_name in edusources_only_fields:
             continue
         # Create the field and its translation
@@ -110,13 +113,14 @@ def migrate_metadata_field_to_entities(apps, schema_editor):
             parents_by_value[value.value] = value
 
 
-def reverse_metadata_field_to_entities():
-    from metadata.models import MetadataField, MetadataTranslation, MetadataValue
-    field_names = [
-        "study_vocabulary.keyword", "disciplines_normalized.keyword", "published_at", "modified_at", "language",
-        "licenses", "technical_types"
-    ]
-    for field_name in field_names:
+def reverse_metadata_field_to_entities(apps, schema_editor):
+    # Skipping this migration for unit test environments
+    if not MetadataField.objects.all().exists():
+        return
+
+    MetadataField.objects.filter(name__in=LEGACY_FIELDS).update(entity="products")
+
+    for field_name in NEW_FIELDS:
         field = MetadataField.objects.get(name=field_name)
         MetadataValue.objects.filter(field__name=field_name).delete()
         MetadataTranslation.objects.filter(metadatavalue__field__name=field_name).delete()
@@ -127,25 +131,12 @@ def reverse_metadata_field_to_entities():
 class Migration(migrations.Migration):
 
     dependencies = [
-        ('metadata', '0006_value_output_order'),
+        ('metadata', '0007_entity'),
     ]
 
     operations = [
-        migrations.AddField(
-            model_name='metadatafield',
-            name='entity',
-            field=models.CharField(choices=[('products', 'products'), ('products:default', 'products:default'), ('products:multilingual-indices', 'products:multilingual-indices'), ('projects', 'projects'), ('projects:default', 'projects:default')], default='products', help_text='Indicates which entity and/or search configuration controls metadata for this field.', max_length=100),
-        ),
-        migrations.AlterUniqueTogether(
-            name='metadatavalue',
-            unique_together={('field', 'value')},
-        ),
-        migrations.RemoveField(
-            model_name='metadatavalue',
-            name='site',
-        ),
         migrations.RunPython(
             migrate_metadata_field_to_entities,
-            migrations.RunPython.noop
+            reverse_metadata_field_to_entities
         )
     ]
